don't write to unmounted /mnt

+++ update.sh
@@ -4,11 +4,13 @@ git pull
 cd ../convictshope
 git pull
 
-sudo rsync * /mnt/
+if [[ -e "/mnt/podium.name" ]]; then
+	sudo rsync * /mnt/
 
-PODIUM=$(cat /mnt/podium.name | sed 's/^hostname: //g')
+	PODIUM=$(cat /mnt/podium.name | sed 's/^hostname: //g')
 
-echo "Podium: $PODIUM"
+	echo "Podium: $PODIUM"
 
-cat config.yml | sed "s/^hostname: .*$/hostname: $PODIUM/g" > /tmp/config.yml
-sudo cp /tmp/config.yml /mnt/config.yml
+	cat config.yml | sed "s/^hostname: .*$/hostname: $PODIUM/g" > /tmp/config.yml
+	sudo cp /tmp/config.yml /mnt/config.yml
+fi
